Started to rework deletion checks so no unintended database issues can arise.
We should really start working with referential integrity, otherwise this will become unmanageable soon.

That would also require things like Taxes be referred to by their IDs, not their names.

https://github.com/FOSSBilling/FOSSBilling/blob/ad94a9f3255801817f393cda36bc0a38796e1eb9/src/install/sql/structure.sql#L575

https://github.com/FOSSBilling/FOSSBilling/blob/ad94a9f3255801817f393cda36bc0a38796e1eb9/src/install/sql/structure.sql#L1365

+++ src/modules/Client/Api/Admin.php
@@ -18,13 +18,13 @@ class Admin extends \Api_Abstract
 {
     /**
      * Get a list of clients.
-     * 
+     *
      * @param array $data Filtering options.
-     * 
+     *
      * @param string $data['status'] [optional] Filter clients by status. Available options: 'active', 'suspended', 'canceled'.
-     * 
+     *
      * @param int $data['per_page'] [optional] Number of clients to display per page.
-     * 
+     *
      * @return array List of clients in a paginated manner.
      */
     public function get_list($data)
@@ -43,11 +43,11 @@ class Admin extends \Api_Abstract
 
     /**
      * Get a list of clients.
-     * 
+     *
      * @param array $data Filtering options
-     * 
+     *
      * @param int $data['per_page'] [optional] Number of clients to display per page.
-     * 
+     *
      * @return array List of clients in a paginated manner
      */
     public function get_pairs($data)
@@ -601,6 +601,14 @@ class Admin extends \Api_Abstract
         $this->di['validator']->checkRequiredParamsForArray($required, $data);
 
         $model = $this->di['db']->getExistingModelById('ClientGroup', $data['id'], 'Group not found');
+
+        $clients = $this->di['db']->find('Client', 'client_group_id = :group_id',[':group_id' => $data['id']]);
+
+        $count = count($clients);
+        if ($count > 0)
+        {
+            throw new \Box_Exception('Group has clients assigned. Please reassign them first.');
+        }
 
         return $this->getService()->deleteGroup($model);
     }
